requirements.txt
streamlit==1.28.1
openai==1.3.0
pdfplumber==0.10.3
PyMuPDF==1.23.8
spacy==3.7.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
reportlab==4.0.7
openpyxl==3.1.2
python-dotenv==1.0.0
linkedin-api==2.2.0
serpapi==0.1.5
nltk==3.8.1
scikit-learn==1.3.2
